Collect timeline marker components in timeline/TracingMarkers.js
