+++ src/components/timeline/TracingMarkers.css
@@ -1,0 +1,9 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+.timelineTracingMarkersCanvas {
+  display: block;
+  width: 100%;
+  height: 100%;
+}
