+++ src/components/timeline/TracingMarkers.js
@@ -9,16 +9,48 @@ import { timeCode } from '../../utils/time-code';
 import { withSize } from '../shared/WithSize';
 import Tooltip from '../shared/Tooltip';
 import MarkerTooltipContents from '../shared/MarkerTooltipContents';
+import {
+  styles,
+  overlayFills,
+} from '../../profile-logic/interval-marker-styles';
+import explicitConnect from '../../utils/connect';
+import { selectorsForThread } from '../../reducers/profile-view';
+import { getSelectedThreadIndex } from '../../reducers/url-state';
+import './TracingMarkers.css';
 
 import type { Milliseconds, CssPixels } from '../../types/units';
 import type { TracingMarker } from '../../types/profile-derived';
 import type { SizeProps } from '../shared/WithSize';
-import type { ConnectedProps } from '../../utils/connect';
+import type {
+  ExplicitConnectOptions,
+  ConnectedProps,
+} from '../../utils/connect';
 import type { ThreadIndex } from '../../types/profile';
 
 type MarkerState = 'PRESSED' | 'HOVERED' | 'NONE';
 
-// Typically this component is wrapped in a connect function, but in other files.
+/**
+ * The TimelineTracingMarkers component is built up of several nested components,
+ * and they are all collected in this file. In pseudo-code, they take
+ * the following forms:
+ *
+ * export const TimelineTracingMarkersJank = (
+ *  <Connect markers={JankMarkers}>
+ *    <WithSize>
+ *      <TimelineTracingMarkers />
+ *    </WithSize>
+ *  </Connect>
+ * );
+ *
+ * export const TimelineTracingMarkersOverview = (
+ *   <Connect markers={AllMarkers}>
+ *     <WithSize>
+ *       <TimelineTracingMarkers />
+ *     </WithSize>
+ *   </Connect>
+ * );
+ */
+
 export type OwnProps = {|
   +className: string,
   +rangeStart: Milliseconds,
@@ -47,7 +79,10 @@ type State = {
   mouseY: CssPixels,
 };
 
-class IntervalMarkerOverview extends React.PureComponent<Props, State> {
+class TimelineTracingMarkersImplementation extends React.PureComponent<
+  Props,
+  State
+> {
   _canvas: HTMLCanvasElement | null;
   _requestedAnimationFrame: boolean | null;
 
@@ -185,18 +220,11 @@ class IntervalMarkerOverview extends React.PureComponent<Props, State> {
 
     const { mouseDownItem, hoveredItem, mouseX, mouseY } = this.state;
     const shouldShowTooltip = !isModifyingSelection && !mouseDownItem;
-    const canvasClassName = className
-      .split(' ')
-      .map(name => `${name}Canvas`)
-      .join(' ');
 
     return (
       <div className={classNames(className, isSelected ? 'selected' : null)}>
         <canvas
-          className={classNames(
-            canvasClassName,
-            'intervalMarkerTimelineCanvas'
-          )}
+          className="timelineTracingMarkersCanvas"
           ref={this._takeCanvasRef}
           onMouseDown={this._onMouseDown}
           onMouseMove={this._onMouseMove}
@@ -323,4 +351,54 @@ class IntervalMarkerOverview extends React.PureComponent<Props, State> {
   }
 }
 
-export default withSize(IntervalMarkerOverview);
+/**
+ * Combine the base implementation of the TimelineTracingMarkers with the
+ * WithSize component.
+ */
+export const TimelineTracingMarkers = withSize(
+  TimelineTracingMarkersImplementation
+);
+
+/**
+ * Create a special connected component for Jank instances.
+ */
+const jankOptions: ExplicitConnectOptions<OwnProps, StateProps, {||}> = {
+  mapStateToProps: (state, props) => {
+    const { threadIndex } = props;
+    const selectors = selectorsForThread(threadIndex);
+    const selectedThread = getSelectedThreadIndex(state);
+
+    return {
+      intervalMarkers: selectors.getJankInstances(state),
+      isSelected: threadIndex === selectedThread,
+      styles: styles,
+      overlayFills: overlayFills,
+    };
+  },
+  component: TimelineTracingMarkers,
+};
+
+export const TimelineTracingMarkersJank = explicitConnect(jankOptions);
+
+/**
+ * Create a connected component for all tracing markers.
+ */
+const tracingOptions: ExplicitConnectOptions<OwnProps, StateProps, {||}> = {
+  mapStateToProps: (state, props) => {
+    const { threadIndex } = props;
+    const selectors = selectorsForThread(threadIndex);
+    const selectedThread = getSelectedThreadIndex(state);
+    const intervalMarkers = selectors.getRangeSelectionFilteredTracingMarkersForHeader(
+      state
+    );
+    return {
+      intervalMarkers,
+      isSelected: threadIndex === selectedThread,
+      styles,
+      overlayFills,
+    };
+  },
+  component: TimelineTracingMarkers,
+};
+
+export const TimelineTracingMarkersOverview = explicitConnect(tracingOptions);
